Move adding rewards to evmos minichef from scripts to deploy folder
